refactor(backend): split learning store memory state and mapper reuse
